Throw error if files are missing.
What was happening if you ran this with a non existent consignment was
it would try to get the files from the api and the api would return
nothing.
This empty list was passed to the s3 downloader which would download
nothing.
The program would continue until we had a bagit package with an empty
data directory.
Arguably, the API should be rejecting non-existent consignments but for
now, this will throw an error if no files are returned by the api.