Run each export in a unique directory
Generate a random folder name for each export. This means that if
the export task is started twice for the same consignment, the two
tasks don't try to read and write the same files in EFS. It also
means that we don't have to clean up the files after running the
export, because a future export of the same consignment will use
a different folder.

This should help to prevent failed exports or corrupted Bagit packages.