Add additional files class
The idea is that all of the additional csv files we need for the bagit
package will be exported from here.

The method first checks each of the option fields in the incoming case
class to see if any are empty, if they are, it fails with an appropriate
message.

Otherwise, it uses the csvWriter library to create the csv.