Replace the existing grapql queries with the consignment metadata one.
We had a query which got the original path for each individual file, we
had a query getting the data for the bagit.txt and another query for
getting the file metadata.

These can all be replaced with the consignment metadata query which cuts
down on network traffic and speeds the task up as well.

This is a big commit but it's mostly removing stuff and refactoring. I
have moved the validation to a separate class though because it was
starting to get scattered everywhere.