Set build sbt to release to github.
This uses a custom task for sbt release. Instead of publishing to s3 or
sonatype, it uses the sbt-github-release plugin to release to github.
It also builds the project when the current release version is set in
version.sbt so we can get the correct version to put in the bag-info
file.
Also, it will generate notes for the release based on the commits since
the last release, but only if no notes have been committed by the dev.